LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-redirect-replace-changed-to-push-expected.txt


FAIL precommitHandler redirect() replace changed to push assert_equals: expected "#redirect1" but got "#replace"
